refactor(examples): expose reusable application roots
Move standalone example setup into library entry points so the showcase can mount the same root widgets and state used by each executable. Keep binary mains as thin shell launchers and gate standalone wasm entry points where imported examples would otherwise export duplicate startup symbols.
