+++ examples/animation-gallery/src/app.rs
@@ -1,7 +1,9 @@
 use crate::chrome;
 use crate::state::AnimationGalleryState;
 use crate::style::BACKGROUND;
-use fission::{build, Container, DesktopApp, Widget};
+#[cfg(not(any(target_arch = "wasm32", target_os = "android", target_os = "ios")))]
+use fission::DesktopApp;
+use fission::{build, Container, Widget};
 
 #[derive(Clone)]
 pub struct AnimationGalleryApp;
@@ -19,6 +21,7 @@ impl From<AnimationGalleryApp> for Widget {
     }
 }
 
+#[cfg(not(any(target_arch = "wasm32", target_os = "android", target_os = "ios")))]
 pub fn run() -> anyhow::Result<()> {
     DesktopApp::<AnimationGalleryState, _>::new(AnimationGalleryApp)
         .with_title("Fission Animation Gallery")
