+++ examples/animation-gallery/src/main.rs
@@ -1,13 +1,3 @@
-mod app;
-mod chrome;
-mod layout;
-mod pages;
-mod routes;
-mod state;
-mod style;
-mod ui;
-mod widgets;
-
 fn main() -> anyhow::Result<()> {
-    app::run()
+    animation_gallery::run()
 }
