+++ examples/editor/Cargo.toml
@@ -8,6 +8,8 @@ fission = { path = "../../crates/authoring/fission", default-features = false, f
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 anyhow = "1.0"
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 lazy_static = "1.4"
 tree-sitter = "0.25"
 tree-sitter-rust = "0.24"
