+++ examples/editor/src/layout.rs
@@ -34,7 +34,9 @@ pub(crate) const SIDEBAR_MAX_WIDTH: f32 = 360.0;
 pub(crate) const MIN_EDITOR_WIDTH: f32 = 180.0;
 pub(crate) const MIN_EDITOR_HEIGHT: f32 = 120.0;
 pub(crate) const MIN_TERMINAL_HEIGHT: f32 = 96.0;
+#[cfg(not(target_arch = "wasm32"))]
 pub(crate) const MIN_TERMINAL_CONTENT_HEIGHT: f32 = 72.0;
+#[cfg(not(target_arch = "wasm32"))]
 pub(crate) const MIN_TERMINAL_WIDTH: f32 = 280.0;
 pub(crate) const TERMINAL_HEIGHT_FRACTION: f32 = 0.33;
 pub(crate) const MINIMAP_WIDTH: f32 = 61.0;
