+++ examples/motion-memory-repro/src/main.rs
@@ -1,14 +1,3 @@
-mod app;
-mod fixtures;
-mod layout;
-mod palette;
-mod repro_row;
-mod repro_screen;
-mod scenario;
-
-use app::MotionMemoryReproApp;
-use fission::prelude::*;
-
 fn main() -> anyhow::Result<()> {
-    DesktopApp::<(), _>::new(MotionMemoryReproApp::from_env()?).run()
+    motion_memory_repro::run_desktop()
 }
