+++ examples/pokemon-card-store/src/lib.rs
@@ -7,5 +7,7 @@ pub mod islands;
 pub mod server;
 pub mod workers;
 
+pub use app::{StoreHomePage, StoreState};
+pub use data::{catalog_response, CatalogRequest, CatalogResponse, StoreError, CATALOG_JOB};
 #[cfg(feature = "server")]
 pub use server::pokemon_card_store_server;
