+++ examples/todo-design-system/src/lib.rs
@@ -1,0 +1,263 @@
+use fission::prelude::*;
+
+include!(concat!(env!("OUT_DIR"), "/todo_design_system.rs"));
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+struct TodoItem {
+    id: usize,
+    title: String,
+    done: bool,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct TodoState {
+    next_id: usize,
+    draft: String,
+    items: Vec<TodoItem>,
+    theme_mode: DesignMode,
+}
+
+impl Default for TodoState {
+    fn default() -> Self {
+        Self {
+            next_id: 4,
+            draft: String::new(),
+            items: vec![
+                TodoItem {
+                    id: 1,
+                    title: "Connect the DSP package".into(),
+                    done: true,
+                },
+                TodoItem {
+                    id: 2,
+                    title: "Use generated component styles".into(),
+                    done: false,
+                },
+                TodoItem {
+                    id: 3,
+                    title: "Ship the themed app".into(),
+                    done: false,
+                },
+            ],
+            theme_mode: DesignMode::Light,
+        }
+    }
+}
+
+impl GlobalState for TodoState {}
+
+#[fission_reducer(UpdateDraft)]
+fn update_draft(state: &mut TodoState, value: String) {
+    state.draft = value;
+}
+
+#[fission_reducer(AddTodo)]
+fn add_todo(state: &mut TodoState) {
+    let title = state.draft.trim();
+    if title.is_empty() {
+        return;
+    }
+    state.items.push(TodoItem {
+        id: state.next_id,
+        title: title.to_string(),
+        done: false,
+    });
+    state.next_id += 1;
+    state.draft.clear();
+}
+
+#[fission_reducer(ToggleTodo)]
+fn toggle_todo(state: &mut TodoState, id: usize) {
+    if let Some(item) = state.items.iter_mut().find(|item| item.id == id) {
+        item.done = !item.done;
+    }
+}
+
+#[fission_reducer(ClearDone)]
+fn clear_done(state: &mut TodoState) {
+    state.items.retain(|item| !item.done);
+}
+
+#[fission_reducer(SetThemeMode)]
+fn set_theme_mode(state: &mut TodoState, mode: DesignMode) {
+    state.theme_mode = mode;
+}
+
+#[derive(Clone)]
+pub struct TodoApp;
+
+impl From<TodoApp> for Widget {
+    fn from(_component: TodoApp) -> Self {
+        let (ctx, view) = fission::build::current::<TodoState>();
+        let colors = &view.env().theme.tokens.colors;
+        let spacing = &view.env().theme.tokens.spacing;
+        let done_count = view.state().items.iter().filter(|item| item.done).count();
+        let add = with_reducer!(ctx, AddTodo, add_todo);
+        let update_draft = with_reducer!(ctx, UpdateDraft(String::new()), update_draft);
+        let clear_done = with_reducer!(ctx, ClearDone, clear_done);
+        let light = with_reducer!(ctx, SetThemeMode(DesignMode::Light), set_theme_mode);
+        let dark = with_reducer!(ctx, SetThemeMode(DesignMode::Dark), set_theme_mode);
+
+        let mut todo_rows = Vec::new();
+        for item in &view.state().items {
+            let toggle = with_reducer!(ctx, ToggleTodo(item.id), toggle_todo);
+            todo_rows.push(
+                Container::new(Row {
+                    gap: Some(12.0),
+                    children: vec![
+                        Checkbox {
+                            checked: item.done,
+                            on_toggle: Some(toggle),
+                            ..Default::default()
+                        }
+                        .into(),
+                        Text::new(item.title.clone())
+                            .size(16.0)
+                            .color(if item.done {
+                                colors.text_secondary
+                            } else {
+                                colors.text_primary
+                            })
+                            .into(),
+                    ],
+                    ..Default::default()
+                })
+                .padding([16.0, 16.0, 12.0, 12.0])
+                .border(colors.border, 1.0)
+                .border_radius(view.env().theme.tokens.radii.medium)
+                .into(),
+            );
+        }
+
+        let theme_toggle = Row {
+            gap: Some(8.0),
+            children: vec![
+                Button {
+                    variant: if view.state().theme_mode == DesignMode::Light {
+                        ButtonVariant::Primary
+                    } else {
+                        ButtonVariant::SecondaryGray
+                    },
+                    size: ComponentSize::Sm,
+                    child: Some(Text::new("Light").into()),
+                    on_press: Some(light),
+                    ..Default::default()
+                }
+                .into(),
+                Button {
+                    variant: if view.state().theme_mode == DesignMode::Dark {
+                        ButtonVariant::Primary
+                    } else {
+                        ButtonVariant::SecondaryGray
+                    },
+                    size: ComponentSize::Sm,
+                    child: Some(Text::new("Dark").into()),
+                    on_press: Some(dark),
+                    ..Default::default()
+                }
+                .into(),
+            ],
+            ..Default::default()
+        }
+        .into();
+
+        let content: Widget = Card {
+            pattern: CardPattern::Raised,
+            child:
+                Column {
+                    gap: Some(20.0),
+                    children: vec![
+                        Row {
+                            gap: Some(12.0),
+                            children: vec![
+                                Column {
+                                    gap: Some(6.0),
+                                    children: vec![
+                                        Text::new("Design-system todo")
+                                            .size(32.0)
+                                            .weight(700)
+                                            .into(),
+                                        Text::new("This example reads a DSP JSON package at build time and uses the generated Rust theme at runtime.")
+                                            .color(colors.text_secondary)
+                                            .into(),
+                                    ],
+                                    flex_grow: 1.0,
+                                    ..Default::default()
+                                }
+                                .into(),
+                                Badge {
+                                    text: format!("{} done", done_count),
+                                    tone: BadgeTone::Success,
+                                    ..Default::default()
+                                }
+                                .into(),
+                            ],
+                            ..Default::default()
+                        }
+                        .into(),
+                        theme_toggle,
+                        Row {
+                            gap: Some(12.0),
+                            children: vec![
+                                TextInput {
+                                    value: view.state().draft.clone(),
+                                    placeholder: Some("Add a task".into()),
+                                    on_change: Some(update_draft),
+                                    width: Some(360.0),
+                                    ..Default::default()
+                                }
+                                .into(),
+                                Button {
+                                    variant: ButtonVariant::Primary,
+                                    child: Some(Text::new("Add task").into()),
+                                    on_press: Some(add),
+                                    ..Default::default()
+                                }
+                                .into(),
+                            ],
+                            ..Default::default()
+                        }
+                        .into(),
+                        Column {
+                            gap: Some(10.0),
+                            children: todo_rows,
+                            ..Default::default()
+                        }
+                        .into(),
+                        Button {
+                            variant: ButtonVariant::TertiaryGray,
+                            child: Some(Text::new("Clear completed").into()),
+                            on_press: Some(clear_done),
+                            disabled: done_count == 0,
+                            ..Default::default()
+                        }
+                        .into(),
+                    ],
+                    ..Default::default()
+                }
+                .into(),
+            ..Default::default()
+        }
+        .into();
+
+        Container::new(content)
+            .bg(colors.background)
+            .padding_all(spacing.xl)
+            .into()
+    }
+}
+
+pub fn configure_embedded_env(state: &TodoState, env: &mut Env) {
+    env.theme = TodoDesignSystem::theme(state.theme_mode);
+}
+
+#[cfg(not(any(target_arch = "wasm32", target_os = "android", target_os = "ios")))]
+pub fn run_desktop() -> anyhow::Result<()> {
+    DesktopApp::<TodoState, _>::new(TodoApp)
+        .with_design_system::<TodoDesignSystem>(DesignMode::Light)
+        .with_sync_env(|state: &TodoState, env: &mut Env| {
+            configure_embedded_env(state, env);
+            env.window.title = WindowTitle::plain("Fission Todo - Design System");
+        })
+        .run()
+}
